Fix silly unintentional constant truncation.
0x100000000 isn't really 0, even if it looks that way
if you truncate it to 32-bit before testing ;)

+++ simplify.c
@@ -202,7 +202,7 @@ static int replace_with_pseudo(struct instruction *insn, pseudo_t pseudo)
 
 static int simplify_constant_rightside(struct instruction *insn)
 {
-	unsigned long value = insn->src2->value;
+	long long value = insn->src2->value;
 
 	switch (insn->opcode) {
 	case OP_ADD: case OP_SUB:
